Initial commit
- callbacks in separate files for more clarity on main app file
